backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

# Data handling and validation
pydantic==2.5.0
requests==2.31.0

# PDF generation
fpdf2==2.7.6

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
hypothesis==6.92.1

# Development dependencies
python-dotenv==1.0.0